Create new KeystreamGenPage with chosen counter if not exists

+++ ChaCha/ChaCha.cs
@@ -327,7 +327,7 @@ namespace Cryptool.Plugins.ChaCha
         public byte[] Xcrypt(byte[] src)
         {
             byte[] dst = new byte[src.Length];
-            int keystreamBlocksNeeded = (int)Math.Ceiling((double)(src.Length) / BLOCKSIZE_BYTES);
+            ulong keystreamBlocksNeeded = (ulong)Math.Ceiling((double)(src.Length) / BLOCKSIZE_BYTES);
             if (!_diffusion)
             {
                 DispatchToPresentation(delegate
@@ -335,7 +335,7 @@ namespace Cryptool.Plugins.ChaCha
                     _presentation.KeystreamBlocksNeeded = keystreamBlocksNeeded;
                 });
             }
-            byte[] keystream = new byte[keystreamBlocksNeeded * BLOCKSIZE_BYTES];
+            byte[] keystream = new byte[keystreamBlocksNeeded * (ulong)BLOCKSIZE_BYTES];
             int keystreamBlocksOffset = 0;
             // Convenience method to abstract away keystream offset.
             void addToKeystream(byte[] block)
